feat: add FlattenColumnsBlock for flattening columns
The `FlattenColumnsBlock` class is added to the `utilblocks.py` file. This block is used to flatten columns in a dataset. It takes an `id_col`, a list of `var_cols`, and a `var_col_name` as input. The `generate` method of this block converts the dataset into a pandas dataframe, performs the flattening operation using the `melt` function, and returns the flattened dataset.

Signed-off-by: shivchander <shivchander.s30@gmail.com>